[Feature]: computer.js, validate.js
컴퓨터가 난수 생성하는 기능, 유저의 입력값 확인하는 기능 모듈화
